Handling of duplicated header text implemented
Common.js has been updated to handle headers with identical text. This
is done by storing every headerId in an array (ids) and for each new id
counting the number of occurences. If the number is larger than 0 (i.e.,
there is at least one other header with the same exact text, then the
number of previous headers is added to the new id after an underscore.

This is part of #21.