Reading charts changed to local files
Previously the monthly reading charts on index.html and
reading/index.html were hosted by Google, and automatically updated
whenever the related spreadsheet was. However Google was using this
opportunity to try to read cookies. Since I don't want a cookie banner,
I have changed this over to local files. This won't make a difference
for the 2023 and 2024 charts since they are no longer updated, but it
does mean that the 2025 chart will have to be manually updated.